+++ compiler/annotation/inference/mod.rs
@@ -1,0 +1,86 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+use std::{
+    borrow::Cow,
+    ops::{Deref, DerefMut},
+};
+
+use std::collections::{BTreeMap, BTreeSet};
+use answer::variable::Variable;
+use ir::pattern::Vertex;
+use answer::Type as TypeAnnotation;
+
+pub(crate) mod match_inference;
+pub(crate) mod type_seeder;
+
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub(crate) struct VertexAnnotations {
+    annotations: BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>>,
+}
+
+impl VertexAnnotations {
+    pub(crate) fn add_or_intersect(
+        &mut self,
+        vertex: &Vertex<Variable>,
+        new_annotations: Cow<'_, BTreeSet<TypeAnnotation>>,
+    ) -> bool {
+        if let Some(existing_annotations) = self.get_mut(vertex) {
+            let size_before = existing_annotations.len();
+            existing_annotations.retain(|x| new_annotations.contains(x));
+            existing_annotations.len() == size_before
+        } else {
+            self.insert(vertex.clone(), new_annotations.into_owned());
+            true
+        }
+    }
+}
+
+impl Deref for VertexAnnotations {
+    type Target = BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.annotations
+    }
+}
+
+impl DerefMut for VertexAnnotations {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.annotations
+    }
+}
+
+impl IntoIterator for VertexAnnotations {
+    type Item = <BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>> as IntoIterator>::Item;
+    type IntoIter = <BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>> as IntoIterator>::IntoIter;
+    fn into_iter(self) -> Self::IntoIter {
+        self.annotations.into_iter()
+    }
+}
+
+impl<'a> IntoIterator for &'a VertexAnnotations {
+    type Item = <&'a BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>> as IntoIterator>::Item;
+    type IntoIter = <&'a BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>> as IntoIterator>::IntoIter;
+    fn into_iter(self) -> Self::IntoIter {
+        self.annotations.iter()
+    }
+}
+
+impl<'a> IntoIterator for &'a mut VertexAnnotations {
+    type Item = <&'a mut BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>> as IntoIterator>::Item;
+    type IntoIter = <&'a mut BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>> as IntoIterator>::IntoIter;
+    fn into_iter(self) -> Self::IntoIter {
+        self.annotations.iter_mut()
+    }
+}
+
+impl<T> From<T> for VertexAnnotations
+where
+    BTreeMap<Vertex<Variable>, BTreeSet<TypeAnnotation>>: From<T>,
+{
+    fn from(t: T) -> Self {
+        Self { annotations: t.into() }
+    }
+}
